Add settings GUI controls to the scene

+++ src/main/java/zagazig/cse/bouncingballs/AppSettings.java
@@ -5,5 +5,5 @@ public class AppSettings {
     public int height;
     public int controlsWidth;
     public int controlsHeight;
-    public int speedMulitplier;
+    public double speedMulitplier;
 }
